# ENH: Speed up _gaussian_kernel1d by removing dependence on np.polynomial