Calculate the kernel of an nth order gaussian filter without using the numpy.polynomial package, which is slow relative to performing the calculations directly.

I ran into a problem where _gaussian_kernel1d was a bottleneck. Avoiding the np.polynomial package drastically speeds things up.

    Old
    ---
    %timeit _gaussian_kernel1d(5, 2, 20)
    432 µs ± 20.9 µs per loop (mean ± std. dev. of 7 runs, 1000 loops each)
 
    New
    ---
    %timeit _gaussian_kernel1d(5, 2, 20)
    30.5 µs ± 3.34 µs per loop (mean ± std. dev. of 7 runs, 10000 loops each)